feat: define SharePlaceIdParser
